Fix JC-556: 1. Rename 'details' folder to 'anime-details' 2. Rename 'table' folder to 'anime-table'
