+++ apps/angular/src/app/features/anime/anime-details/anime-details.component.css
