+++ apps/angular/src/app/features/anime/anime-routing.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AuthGuard } from '@js-camp/angular/core/services/auth-guard.service';
 
-import { AnimeTableComponent } from './table/anime-table.component';
-import { AnimeDetailsComponent } from './details/anime-details.component';
+import { AnimeTableComponent } from './anime-table/anime-table.component';
+import { AnimeDetailsComponent } from './anime-details/anime-details.component';
 
 const routes: Routes = [
 	{
