+++ apps/angular/src/app/features/anime/anime-table/anime-table.component.html
