+++ apps/angular/src/app/features/anime/anime.module.ts
@@ -14,9 +14,9 @@ import { MatCardModule } from '@angular/material/card';
 import { SafePipe } from '@js-camp/angular/core/services/dom-sanitizer.service';
 import { DialogModule } from '@angular/cdk/dialog';
 
-import { AnimeTableComponent } from './table/anime-table.component';
+import { AnimeTableComponent } from './anime-table/anime-table.component';
 import { AnimeRoutingModule } from './anime-routing.module';
-import { AnimeDetailsComponent } from './details/anime-details.component';
+import { AnimeDetailsComponent } from './anime-details/anime-details.component';
 import { PosterPopupComponent } from './components/poster-popup/poster-popup.component';
 import { HeaderComponent } from './components/header/header.component';
